% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_align.R
\name{txm_align}
\alias{txm_align}
\title{Customized local BLAST}
\usage{
txm_align(
  seq_in,
  db_path = NULL,
  db_name = NULL,
  tab_out = paste("Output", Sys.Date(), sep = ""),
  tab_path = ".",
  acsn_list = NULL,
  acsn_path = ".",
  alt_path = NULL,
  alt_annot = T,
  org = "bac",
  asbd_tbl_lge = NA,
  asgn_tbl_lge = paste("Dataset_lge", Sys.Date(), ".fst", sep = ""),
  task = "megablast",
  threads = 1,
  acsn_check = F,
  show = F,
  run_blst = T,
  batches = 1,
  chunks = 1,
  qcvg = 98,
  pctidt = 98,
  max_out = 500,
  min_hits = 1
)
}
\arguments{
\item{seq_in}{\strong{Required} All input types accepted by
\link[dada2]{getSequences}.}

\item{db_path}{(String) \strong{Required} Full path for local BLAST database being used for
alignment.}

\item{db_name}{(String) \strong{Required} Name of BLAST database.}

\item{tab_out}{(String) Default "Output" + System date. Specify name of
the output files.}

\item{tab_path}{(String) Default "." (current working directory). Specify
full output path for results.}

\item{acsn_list}{(String) Default NULL. Full path to a list of accession
IDs that will be used to restrict the BLAST database (-seqidlist), which
is highly recommended for large databases such as "nt/nr". This can be
obtained using the \link[taxminer]{txm_accIDs} function provided within this
package. Further information regarding search limitations is available on
\href{https://www.ncbi.nlm.nih.gov/books/NBK279673/}{BLAST user manual}}

\item{acsn_path}{(String) Default "." (current working directory). Specify
full path to the accession ID list.}

\item{alt_path}{(Required) Default NULL. Path to SILVA and RDP databases.
Must identify a folder that contains at least 4 files:
\itemize{
\item silva training set (regex: "silva.*train")
\item silva species assignment (regex: "silva.*species")
\item rdp train set (regex: "rdp.*train")
\item rdp species assignment (regex: "rdp.*species")
\item UNITE general release fasta (regex: "sh_general.*fasta")
}
Pre-formatted dada2 databases are recommended:
\href{https://benjjneb.github.io/dada2/training.html}{dada2 fasta files}}

\item{alt_annot}{(Logical) Default TRUE. Alternative annotations with SILVA
and RDP databases using \link[dada2]{assignTaxonomy} for up to genus level
annotations followed by \link[dada2]{addSpecies} for assigning species.}

\item{org}{(String) Default "bac". The organism being investigated.
\itemize{
\item "bac" - Bacteria. Silva and RDP databases are used for alternative
annotations to BLAST.
\item "fungi" - Fungi. UNITE database is used for alternative annotations
to BLAST.
}}

\item{asbd_tbl_lge}{(String) Default NA. Full path and name of a
pre-compiled table to be used for lineage assignment}

\item{asgn_tbl_lge}{(String) Default Dataset_lge + system date. Full path
and name of a new data set to be created.}

\item{task}{(String) Default "megablast". Other possible tasks can be
found here:
\href{https://www.ncbi.nlm.nih.gov/books/NBK569839/#usrman_BLAST_feat.Tasks}{Tasks}}

\item{threads}{(>0) Default 1. Number of threads assigned to BLAST
alignment. Set these in accordance to your hardware.}

\item{acsn_check}{(Logical) Default FALSE. If an accession ID list is
provided, BLASTDB v5 requires it to be pre-processed (blastdb_aliastool)
prior to being used for restricting the database. Set this to TRUE if an
unprocessed accession ID list is specified.}

\item{show}{(Logical) Default FALSE. Switch from console to terminal?}

\item{run_blst}{(Logical) Default TRUE. Run BLAST alignment?}

\item{batches}{(>0) Default 1. Should the sequences be aligned in batches?
\itemize{
\item 1 = no batches
\item >1 = total sequences/batches
}}

\item{chunks}{(>0) Default 1. Should each batch of sequences be split into
further chunks to be aligned in parallel?
\itemize{
\item 1 = no further chunks
\item >1 = each batch is further divided to the closest number of chunks
possible and ran in parallel using GNU parallel
}}

\item{qcvg}{(0-100) Default 98 (\%). Threshold for query coverage.}

\item{pctidt}{(0-100) Default 98 (\%). Threshold for percentage identity.}

\item{max_out}{(>0) Default 500. Maximum number of alignments
("-max_target_seqs"). Higher values are recommended when using large
databases such as "nt/nr".}

\item{min_hits}{(Whole number) Default 1. Minimum number of hits per sequence
alignment. IDs with hits lower than this threshold will be removed.}
}
\description{
Build and run a customized BLAST alignment using local cmd line BLAST
and locally downloaded databases from within RStudio. User defined seq_ins
and specifications are used to build a BLAST cmd, which is transferred
to RStudio terminal. The output is formatted as BLASTn tabular output
format 6.
In parallel, SILVA and RDP databases will be used to assign alternative
annotations that are further used for annotations scores.
}
