% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_ecosrc.R
\name{txm_ecosrc}
\alias{txm_ecosrc}
\title{Text mining and filtration}
\usage{
txm_ecosrc(
  hit_tbl,
  filt_host = NA,
  filt_site = NA,
  filt_negt = NA,
  alt_tbl_path = NULL,
  org = "bac",
  do_filt = T,
  asbd_tbl = NA,
  asgn_tbl = paste("Dataset_", Sys.Date(), ".fst", sep = ""),
  sys.break = 1
)
}
\arguments{
\item{hit_tbl}{(data.frame) \strong{Required} Default NULL. Output table obtained
from \link[taxminer]{txm_align}.}

\item{filt_host}{(String) Default NA. Filter annotations by host}

\item{filt_site}{(String) Default NA. Filter annotations by body site or
environment.}

\item{filt_negt}{(String) Default NA. Disregard annotations that contain
these terms.}

\item{alt_tbl_path}{(String) \strong{Required} Default NULL. Full path to silva
and RDP annotation outputs.}

\item{org}{(String) Default "bac". The organism being investigated.
\itemize{
\item "bac" - Bacteria. Silva and RDP databases are used for alternative
annotations to BLAST.
\item "fungi" - Fungi. UNITE database is used for alternative annotations
to BLAST.
}}

\item{do_filt}{(Logical) Default TRUE. Perform filtration using the word
banks. If FALSE the output will contain all accession IDs and the extracted
information associated to them.}

\item{asbd_tbl}{(String) Default NA. Specify the name of the pre-compiled
database present within the directory.}

\item{asgn_tbl}{(String) Default Dataset + system date. Name of a new
compiled database to be assigned}

\item{sys.break}{(>0) Default 1. Amount of time, in seconds, that the
system is paused between iterations. This is handy with larger
queries to reduce the load on the NCBI servers.}
}
\description{
Extract information from NCBI nucleotide and PubMed databases, attaching
ecosystem specificity to each accession ID. Different combinations of word
banks are used to scan through the data and apply the filtration
criteria. The accession IDs are split into batches of 200,
and \link[rentrez]{rentrez} is used to communicate with NCBI.
For each annotation hit a "score" is calculated based on the consensus
between the superkingdom-species level lineage obtained from
BLAST + Silva + RDP and the completeness of the available
ecosystem/publication information.
Uses a NCBI api key is highly recommended.
\href{  https://cran.r-project.org/web/packages/rentrez/vignettes/rentrez_tutorial.html#rate-limiting-and-api-keys}{rentrez vignette}
}
