% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_lineage.R
\name{txm_lineage}
\alias{txm_lineage}
\title{Extract lineage}
\usage{
txm_lineage(
  taxids,
  asbd_tbl = NA,
  asgn_tbl = paste("Dataset_lge", Sys.Date(), ".fst", sep = "")
)
}
\arguments{
\item{taxids}{(data.frame) \strong{Required} Output table obtained
from \link[taxminer]{txm_ecosrc}. Alternatively, a data.frame with a single
column named "TaxID" and optionally a second column named "AccID".}

\item{asbd_tbl}{(String) Default NA. Specify the name of the
pre-assembled database present within the directory.}

\item{asgn_tbl}{(String) Default Dataset_lge + system date. Name of a new
database to be assembled}
}
\description{
Obtain full lineage from taxonomic IDs. Splits TaxIDs into batches of 200 and uses
\link[rentrez:entrez_post]{entrez_post} to post them onto the NCBI
web history server, followed by using
\link[rentrez:entrez_fetch]{entrez_fetch} to obtain xml files.
}
