% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunc.R
\docType{data}
\name{Word_banks}
\alias{Word_banks}
\title{Word banks}
\format{
An object of class \code{list} of length 8.
}
\usage{
Word_banks
}
\description{
Full list of word banks being used for text mining based filtration
}
\keyword{datasets}
