% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_align.R
\name{txm_align}
\alias{txm_align}
\title{Customized local BLAST}
\usage{
txm_align(
  seq_in,
  db_path = NULL,
  db_name = NULL,
  tab_out = paste("Output", Sys.Date(), sep = ""),
  tab_path = ".",
  acsn_list = NULL,
  acsn_path = ".",
  task = "megablast",
  threads = 1,
  acsn_check = F,
  show = F,
  run_blst = T,
  qcvg = 98,
  pctidt = 98,
  max_out = 500
)
}
\arguments{
\item{seq_in}{(Required) All seq_in types accepted by
\link[dada2]{getSequences}.}

\item{db_path}{(Required) Full path for local BLAST database being used for
alignment.}

\item{db_name}{(Required) Name of BLAST database.}

\item{tab_out}{(Optional) Default "Output" + System date. Specify name of
the output files.}

\item{tab_path}{(Optional) Default "." (current working directory). Specify
full output path for results.}

\item{acsn_list}{(Optional) Default NULL. Full path to a list of accession
IDs that will be used to restrict the BLAST database (-seqidlist), which
is highly recommended for large databases such as "nt/nr". This can be
obtained using the \link[taxminer]{txm_accIDs} function provided within this
package. Further information regarding search limitations is available on
\href{https://www.ncbi.nlm.nih.gov/books/NBK279673/}{BLAST user manual}}

\item{acsn_path}{(Optional) Default "." (current working directory). Specify
full path to the accession ID list.}

\item{task}{(Optional) Default "megablast". Other possible tasks can be
found here:
\href{https://www.ncbi.nlm.nih.gov/books/NBK569839/#usrman_BLAST_feat.Tasks}{Tasks}}

\item{threads}{(Optional) Default 1. Number of threads assigned to BLAST
alignment.}

\item{acsn_check}{(Logical) Default FALSE. If an accession ID list is
provided, BLASTDB v5 requires it to be pre-processed (blastdb_aliastool)
prior to being used for restricting the database. Set this to TRUE if an
unprocessed accession ID list is specified.}

\item{show}{(Logical) Default FALSE. Switch from console to terminal?}

\item{run_blst}{(Logical) Default TRUE. Set to FALSE if an existing
comma-delimited alignment file is present in the directory.}

\item{qcvg}{(Optional) Default 98 (\%). Threshold for query coverage.}

\item{pctidt}{(Optional) Default 98 (\%). Threshold for percentage identity.}

\item{max_out}{(Optional) Default 500. Maximum number of alignments
("-max_target_seqs"). Higher values are recommended when using large
databases such as "nt/nr".}
}
\description{
Build and run a customized BLAST alignment using local cmd line BLAST
and locally downloaded databases from within RStudio. User defined seq_ins
and specifications are used to build a BLAST cmd, which is transferred
to RStudio terminal. The output is formatted as BLASTn tabular output
format 6.
}
