% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txm_align.R
\name{txm_align}
\alias{txm_align}
\title{Customized local BLAST}
\usage{
txm_align(
  input,
  previous_run = NULL,
  task = "megablast",
  database_path = "~/Documents/NCBI_databases/16S",
  database_name = "16S_ribosomal_RNA",
  output_name = paste("Output", Sys.Date(), sep = ""),
  threads = 1,
  accession_list = "Bacteria_noEnv.seq",
  do_acc_check = F,
  show = F,
  Run_Blast = T,
  qcvg = 98,
  pctidt = 98,
  max_out = 500
)
}
\arguments{
\item{input}{(Required) All input types accepted by \link[dada2]{getSequences}.}

\item{previous_run}{(Optional) Default NULL. Remove previously aligned sequences
(must contain valid IDs)}

\item{task}{(Optional) Default "megablast".}

\item{database_path, database_name}{(Required) Full path and name for a local BLAST
database being used for alignment.}

\item{output_name}{(Optional) Default "Output" + System date.
Specify name of the output files.}

\item{threads}{(Optional) Default 1. Number of threads assigned to BLAST alignment.}

\item{accession_list}{(Optional) Default NULL. A list of accession IDs that will be
used to restrict the BLAST database (-seqidlist), which is highly recommended
for large databases such as "nt/nr". This can be obtained using
the \link[taxminer]{txm_accIDs} function provided within this
package. Further information regarding search limitations is
available on
\href{https://www.ncbi.nlm.nih.gov/books/NBK279673/}{BLAST command line user manual}}

\item{do_acc_check}{(Logical) Default FALSE. If an accession ID list is provided, BLASTDB v5 requires it to be
pre-processed (blastdb_aliastool) prior to being used for restricting the database.
Set this to TRUE is an unprocessed accession ID list is specified.}

\item{show}{(Logical) Default FALSE. Switch from console to terminal?}

\item{Run_Blast}{(Logical) Default TRUE. Set to FALSE if an existing comma-delimited alignment file is present
in the directory and the function should be utilized to process it.}

\item{qcvg}{(Optional) Default 98. Query coverage filtration in alignment as a percentage.}

\item{pctidt}{(Optional) Default 98. Percentage identity filtration in alignment as a percentage}

\item{max_out}{(Optional) Default 500. Maximum number of alignments ("-max_target_seqs"). This should be set much
higher when using large databases such as "nt/nr", especially when other restrictions are kept to a
minimum.}
}
\description{
Build and run a customized BLAST alignment using local command line BLAST and locally downloaded databases from
within RStudio. User defined inputs and specifications are used to build a BLAST command, which is
transferred to RStudio terminal. The output is formatted as BLASTn tabular output format 6.
}
